requirements.txt
Django==3.2
psycopg2-binary==2.9.1
pandas==1.3.1
djangorestframework==3.12.2
django-restql==0.5.3
pytest-django==4.4.0
pytest-mock==3.6.1
coverage==5.5